requirements.txt
affine==2.4.0
asgiref==3.7.2
attrs==25.3.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1.2
cligj==0.7.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
Django==5.0.1
django-apps==0.1.0
django-test==0.4030
django_db==0.0.7
django_urls==1.1.3
djangorestframework==3.15.1
et-xmlfile==1.1.0
fire==0.6.0
Flask==3.0.2
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
fonttools==4.53.1
greenlet==3.0.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mysql-connector-python==8.3.0
mysqlclient==2.2.4
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.10.0.84
openpyxl==3.1.2
packaging==24.1
pandas==2.1.4
pdf2docx==0.5.8
pillow==10.4.0
pygame==2.6.0
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
PyMySQL==1.1.0
pyparsing==3.1.2
pypdf==4.3.1
PyPDF2==3.0.1
python-dateutil==2.8.2
python-docx==1.1.2
pytube==15.0.0
pytz==2023.3.post1
rasterio==1.4.3
requests==2.31.0
setuptools==69.0.3
six==1.16.0
SQLAlchemy==2.0.28
sqlparse==0.4.4
termcolor==2.4.0
typing_extensions==4.10.0
tzdata==2023.4
urllib3==2.1.0
Werkzeug==3.0.1
